service_auth/requirements.txt
backoff==2.2.1
gevent==22.10.2
gunicorn==20.1.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
Flask-Limiter==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-RESTful==0.3.9
flasgger==0.9.5
psycopg2==2.9.5
pydantic==1.10.2
redis==4.4.0
SQLAlchemy==1.4.45
python-dotenv==0.21.0
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-exporter-jaeger==1.15.0
rauth==0.7.3
